MP-58 :truck: Rename: Redis 파일 및 디렉토리 수정
- 기존 utils 클래스들과의 혼동을 방지하기 위해 RedisUtil을 RedisHelper로 이름 변경
- RedisConfig와 RedisHelper를 global.middleware.redis로 이동
